First working config
This commits marks the beginning of my emacs journey. Right now, I
have a somewhat nicely working config with a few basic settings, much
more is to be added in the future. I will be trying to keep this
configuration as mininal as feasible and convenient, as I've come to
dislike bloated and slow configurations, which are just not my thing.